+++ fe/fe-core/src/test/java/com/starrocks/alter/CompactionHandlerTest.java
@@ -12,26 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// This file is based on code available under the Apache license here:
-//   https://github.com/apache/incubator-doris/blob/master/fe/fe-core/src/test/java/org/apache/doris/alter/AlterTest.java
-
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
 package com.starrocks.alter;
 
 import com.starrocks.catalog.Database;
@@ -48,8 +28,6 @@ import com.starrocks.sql.ast.CompactionClause;
 import com.starrocks.utframe.StarRocksAssert;
 import com.starrocks.utframe.UtFrameUtils;
 import mockit.Expectations;
-import mockit.Mock;
-import mockit.MockUp;
 import mockit.Mocked;
 import org.junit.Assert;
 import org.junit.BeforeClass;
@@ -61,10 +39,6 @@ import java.util.List;
 
 import static org.mockito.Mockito.mock;
 
-/**
- * @author xiaolong
- * @date 2024/02/06/15:56
- */
 public class CompactionHandlerTest {
 
     private static Database db;
@@ -100,13 +74,6 @@ public class CompactionHandlerTest {
                 ")\n" +
                 "DISTRIBUTED BY HASH(v1) BUCKETS 3\n" +
                 "PROPERTIES('replication_num' = '1');");
-
-        new MockUp<RunMode>() {
-            @Mock
-            public RunMode getCurrentRunMode() {
-                return RunMode.SHARED_DATA;
-            }
-        };
 
         db = GlobalStateMgr.getCurrentState().getDb(GlobalStateMgrTestUtil.testDb1);
         olapTable = (OlapTable) db.getTable(GlobalStateMgrTestUtil.testTable1);
